changed review responses to floats

+++ movie-reviews/dtos/responses.py
@@ -2,4 +2,4 @@ from typing import List
 from pydantic import BaseModel
 
 class PredictResponse(BaseModel):
-    ratings: List[int]
+    ratings: List[float]
